+++ components/calendar/src/indian.rs
@@ -376,7 +376,7 @@ mod tests {
         iso_day: u8,
         expected_year: i32,
         expected_month: u8,
-        expected_day: u32,
+        expected_day: u8,
     }
 
     fn check_case(case: TestCase) {
